refactor(constant): move enum file to constant directory
PriceRule
CompareInteger
